backend/requirements.txt
Django>=5.1,<5.2
django-modeltranslation>=0.18,<1.0
djangorestframework>=3.15,<3.16
django-cors-headers>=4.3,<4.4
django-unfold>=0.40,<1.0
psycopg[binary]>=3.2,<3.3
Pillow>=10.0,<11.0
gunicorn>=22.0,<23.0
whitenoise>=6.7,<7.0
dj-database-url>=2.1,<2.2
python-decouple>=3.8,<3.9
